Don't expose broken or fragile APIs to plugins
LP: #1571264
