+++ bin/snapcraft
@@ -26,6 +26,6 @@ if os.path.exists(os.path.join(topdir, 'setup.py')):
 
 if __name__ == '__main__':
     import snapcraft.main
-    import snapcraft.dirs
-    snapcraft.dirs.setup_dirs()
+    import snapcraft.internal.dirs
+    snapcraft.internal.dirs.setup_dirs()
     snapcraft.main.main()
