+++ snapcraft/internal/__init__.py
@@ -14,4 +14,5 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from snapcraft.internal import states  # noqa
+from snapcraft.internal import states            # noqa
+from snapcraft.internal.yaml import load_config  # noqa
