+++ snapcraft/internal/log.py
