+++ snapcraft/tests/test_base_plugin.py
@@ -19,11 +19,8 @@ import tempfile
 import unittest.mock
 
 import snapcraft
-from snapcraft import (
-    common,
-    sources,
-    tests
-)
+from snapcraft.internal import common
+from snapcraft import sources, tests
 
 
 class TestBasePlugin(tests.TestCase):
