+++ snapcraft/tests/test_commands_clean.py
@@ -20,11 +20,11 @@ import shutil
 from unittest import mock
 
 from snapcraft.main import main
-from snapcraft import (
-    internal,
+from snapcraft.internal import (
     pluginhandler,
-    tests,
+    states,
 )
+from snapcraft import tests
 
 
 class CleanCommandTestCase(tests.TestCase):
@@ -152,15 +152,15 @@ parts:
         main(['clean', '--step=foo'])
 
         expected_staged_state = {
-            'clean0': internal.states.StageState({'clean0'}, set()),
-            'clean1': internal.states.StageState({'clean1'}, set()),
-            'clean2': internal.states.StageState({'clean2'}, set()),
+            'clean0': states.StageState({'clean0'}, set()),
+            'clean1': states.StageState({'clean1'}, set()),
+            'clean2': states.StageState({'clean2'}, set()),
         }
 
         expected_stripped_state = {
-            'clean0': internal.states.StripState({'clean0'}, set()),
-            'clean1': internal.states.StripState({'clean1'}, set()),
-            'clean2': internal.states.StripState({'clean2'}, set()),
+            'clean0': states.StripState({'clean0'}, set()),
+            'clean1': states.StripState({'clean1'}, set()),
+            'clean2': states.StripState({'clean2'}, set()),
         }
 
         mock_clean.assert_called_with(
