+++ snapcraft/tests/test_plugin_go.py
@@ -30,7 +30,7 @@ class GoPluginTestCase(tests.TestCase):
 
         self.project_options = snapcraft.ProjectOptions()
 
-        patcher = mock.patch('snapcraft.common.run')
+        patcher = mock.patch('snapcraft.internal.common.run')
         self.run_mock = patcher.start()
         self.addCleanup(patcher.stop)
